Create git_push_automate.sh
To use this script, you'll need to replace /path/to/repository with the path to your Git repository. Also, modify the commit_message variable to your desired commit message.

The script navigates to the specified repository directory using cd. It then uses git add -A to stage all modified and new files, followed by git commit to commit the changes with the specified commit message. Finally, it uses git push to push the changes to the default branch of the repository.

Please note that you should ensure you have the necessary permissions and have properly configured your Git settings (e.g., remote repository and default branch) for the script to work correctly.