Create log_cleaner.sh
To use this script, you'll need to replace /path/to/logs with the path to the directory containing the log files you want to clean. You can also adjust the days_to_keep variable to define how many days' worth of logs to keep.

Save the script to a file, such as log_cleaner.sh, and make it executable.

The script iterates over the log files in the specified directory. It calculates the cutoff date based on the number of days to keep logs and compares it with the last modified date of each file. If a file's last modified date is older than the cutoff date, it gets removed using rm. The script also prints a message for each removed file.

After running the script, the log files in the specified directory that are older than the defined number of days will be removed, freeing up disk space.