Create outputs.tf
outputs.tf file declares a single output variable named rds_instance_id. Here's an overview of this output:

The purpose of this output declaration is to expose the ID of the created RDS instance as an output value that can be accessed and utilized by other parts of your Terraform configuration, such as the root module or other modules that consume the rds_instance_id output variable.